Moving Average Trader implementing its own indicators TODO : Indicators should use duration for their periods

+++ ts/src/main/scala/ch/epfl/ts/indicators/SmaIndicator.scala
@@ -1,17 +1,17 @@
 package ch.epfl.ts.indicators
 
-case class SMA(override val value: Map[Int, Double]) extends MovingAverage(value)
+case class SMA(override val value: Map[Long, Double]) extends MovingAverage(value)
 
 /**
  * Simple moving average indicator
  */
-class SmaIndicator(periods: List[Int]) extends MaIndicator(periods) {
+class SmaIndicator(periods: List[Long]) extends MaIndicator(periods) {
 
   def computeMa: SMA = {
 
-    def auxCompute(period: Int): Double = {
+    def auxCompute(period: Long): Double = {
       var sma: Double = 0.0
-      values.takeRight(period).map { o => sma = sma + o.close }
+      values.takeRight(period.toInt).map { o => sma = sma + o.close }
       sma / period
     }
     SMA(periods.map(p => (p -> auxCompute(p))).toMap)
